+++ src/main/java/net/sf/jabref/gui/JabRefFrame.java
@@ -124,6 +124,7 @@ import net.sf.jabref.model.database.DatabaseLocation;
 import net.sf.jabref.model.entry.BibEntry;
 import net.sf.jabref.model.entry.BibtexEntryTypes;
 import net.sf.jabref.model.entry.EntryType;
+import net.sf.jabref.model.entry.FieldName;
 import net.sf.jabref.preferences.HighlightMatchingGroupPreferences;
 import net.sf.jabref.preferences.JabRefPreferences;
 import net.sf.jabref.preferences.LastFocusedTabPreferences;
@@ -302,8 +303,6 @@ public class JabRefFrame extends JFrame implements OutputPrinter {
             Printed.getInstance().getValues().get(0).getMenuString(),
             Printed.getInstance().getValues().get(0).getToolTipText(),
             IconTheme.JabRefIcon.PRINTED.getIcon());
-    private final AbstractAction manageSelectors = new GeneralAction(Actions.MANAGE_SELECTORS,
-            Localization.menuTitle("Manage content selectors"));
     private final AbstractAction normalSearch = new GeneralAction(Actions.SEARCH, Localization.menuTitle("Search"),
             Localization.lang("Search"), Globals.getKeyPrefs().getKey(KeyBinding.SEARCH), IconTheme.JabRefIcon.SEARCH.getIcon());
     private final AbstractAction globalSearch = new GeneralAction(Actions.GLOBAL_SEARCH,
@@ -504,6 +503,10 @@ public class JabRefFrame extends JFrame implements OutputPrinter {
     private final List<Object> openAndSavedDatabasesOnlyActions = new LinkedList<>();
     private final List<Object> sharedDatabaseOnlyActions = new LinkedList<>();
     private final List<Object> noSharedDatabaseActions = new LinkedList<>();
+    private final List<Object> oneEntryOnlyActions = new LinkedList<>();
+    private final List<Object> oneEntryWithFileOnlyActions = new LinkedList<>();
+    private final List<Object> oneEntryWithURLorDOIOnlyActions = new LinkedList<>();
+    private final List<Object> twoEntriesOnlyActions = new LinkedList<>();
 
 
     private class EditModeAction extends AbstractAction {
@@ -587,10 +590,6 @@ public class JabRefFrame extends JFrame implements OutputPrinter {
         JMenuItem bibtexKeyPatternBtn = new JMenuItem(Localization.lang("BibTeX key patterns"));
         bibtexKeyPatternBtn.addActionListener(bibtexKeyPattern);
         popupMenu.add(bibtexKeyPatternBtn);
-
-        JMenuItem manageSelectorsBtn = new JMenuItem(Localization.lang("Manage content selectors"));
-        manageSelectorsBtn.addActionListener(manageSelectors);
-        popupMenu.add(manageSelectorsBtn);
 
         return popupMenu;
     }
@@ -1336,7 +1335,6 @@ public class JabRefFrame extends JFrame implements OutputPrinter {
         options.add(customImpAction);
         options.add(customFileTypesAction);
         options.add(manageJournals);
-        options.add(manageSelectors);
         options.add(protectTerms);
         options.add(selectKeys);
         mb.add(options);
@@ -1504,7 +1502,7 @@ public class JabRefFrame extends JFrame implements OutputPrinter {
 
     private void initActions() {
         openDatabaseOnlyActions.clear();
-        openDatabaseOnlyActions.addAll(Arrays.asList(manageSelectors, mergeDatabaseAction, newSubDatabaseAction, save, globalSearch,
+        openDatabaseOnlyActions.addAll(Arrays.asList(mergeDatabaseAction, newSubDatabaseAction, save, globalSearch,
                 saveAs, saveSelectedAs, saveSelectedAsPlain, editModeAction, undo, redo, cut, deleteEntry, copy, paste, mark, markSpecific, unmark,
                 unmarkAll, rankSubMenu, editEntry, selectAll, copyKey, copyCiteKey, copyKeyAndTitle, editPreamble, editStrings,
                 toggleGroups, makeKeyAction, normalSearch, generalFetcher.getAction(), mergeEntries, cleanupEntries, exportToClipboard, replaceAll,
@@ -1527,6 +1525,18 @@ public class JabRefFrame extends JFrame implements OutputPrinter {
         openAndSavedDatabasesOnlyActions.addAll(Collections.singletonList(openConsole));
         sharedDatabaseOnlyActions.addAll(Collections.singletonList(pullChangesFromSharedDatabase));
         noSharedDatabaseActions.addAll(Arrays.asList(save, saveAll));
+
+        oneEntryOnlyActions.clear();
+        oneEntryOnlyActions.addAll(Arrays.asList(editEntry));
+
+        oneEntryWithFileOnlyActions.clear();
+        oneEntryWithFileOnlyActions.addAll(Arrays.asList(openFolder, openFile));
+
+        oneEntryWithURLorDOIOnlyActions.clear();
+        oneEntryWithURLorDOIOnlyActions.addAll(Arrays.asList(openUrl));
+
+        twoEntriesOnlyActions.clear();
+        twoEntriesOnlyActions.addAll(Arrays.asList(mergeEntries));
 
         tabbedPane.addChangeListener(event -> updateEnabledState());
 
@@ -1583,6 +1593,14 @@ public class JabRefFrame extends JFrame implements OutputPrinter {
             boolean isShared = current.getBibDatabaseContext().getLocation() == DatabaseLocation.SHARED;
             setEnabled(sharedDatabaseOnlyActions, isShared);
             setEnabled(noSharedDatabaseActions, !isShared);
+
+            boolean oneEntrySelected = current.getSelectedEntries().size() == 1;
+            setEnabled(oneEntryOnlyActions, oneEntrySelected);
+            setEnabled(oneEntryWithFileOnlyActions, isExistFile(current.getSelectedEntries()));
+            setEnabled(oneEntryWithURLorDOIOnlyActions, isExistURLorDOI(current.getSelectedEntries()));
+
+            boolean twoEntriesSelected = current.getSelectedEntries().size() == 2;
+            setEnabled(twoEntriesOnlyActions, twoEntriesSelected);
         }
     }
 
@@ -1883,6 +1901,34 @@ public class JabRefFrame extends JFrame implements OutputPrinter {
             SwingUtilities.invokeLater(() -> progressBar.setMaximum(value));
         }
 
+    }
+
+    /**
+     * Return a boolean, if the selected entry have file
+     * @param selectEntryList A selected entries list of the current base pane
+     * @return true, if the selected entry contains file.
+     * false, if multiple entries are selected or the selected entry doesn't contains file
+     */
+    private boolean isExistFile(List<BibEntry> selectEntryList) {
+        if (selectEntryList.size() == 1) {
+            BibEntry selectedEntry = selectEntryList.get(0);
+            return selectedEntry.getField(FieldName.FILE).isPresent();
+        }
+        return false;
+    }
+
+    /**
+     * Return a boolean, if the selected entry have url or doi
+     * @param selectEntryList A selected entries list of the current base pane
+     * @return true, if the selected entry contains url or doi.
+     * false, if multiple entries are selected or the selected entry doesn't contains url or doi
+     */
+    private boolean isExistURLorDOI(List<BibEntry> selectEntryList) {
+        if (selectEntryList.size() == 1) {
+            BibEntry selectedEntry = selectEntryList.get(0);
+            return (selectedEntry.getField(FieldName.URL).isPresent() || selectedEntry.getField(FieldName.DOI).isPresent());
+        }
+        return false;
     }
 
     private class ChangeTabAction extends MnemonicAwareAction {