+++ src/main/java/net/sf/jabref/gui/importer/actions/AppendDatabaseAction.java
@@ -71,13 +71,12 @@ public class AppendDatabaseAction implements BaseAction {
             // Run the actual open in a thread to prevent the program
             // locking until the file is loaded.
             JabRefExecutorService.INSTANCE.execute(
-                    () -> openIt(md.importEntries(), md.importStrings(), md.importGroups(), md.importSelectorWords()));
+                    () -> openIt(md.importEntries(), md.importStrings(), md.importGroups()));
         }
 
     }
 
-    private void openIt(boolean importEntries, boolean importStrings, boolean importGroups,
-            boolean importSelectorWords) {
+    private void openIt(boolean importEntries, boolean importStrings, boolean importGroups) {
         if (filesToOpen.isEmpty()) {
             return;
         }
@@ -87,8 +86,7 @@ public class AppendDatabaseAction implements BaseAction {
                 // Should this be done _after_ we know it was successfully opened?
                 ParserResult pr = OpenDatabase.loadDatabase(file,
                         Globals.prefs.getImportFormatPreferences());
-                AppendDatabaseAction.mergeFromBibtex(frame, panel, pr, importEntries, importStrings,
-                        importGroups, importSelectorWords);
+                AppendDatabaseAction.mergeFromBibtex(frame, panel, pr, importEntries, importStrings, importGroups);
                 panel.output(Localization.lang("Imported from database") + " '" + file.getPath() + "'");
             } catch (IOException | KeyCollisionException ex) {
                 LOGGER.warn("Could not open database", ex);
@@ -99,7 +97,7 @@ public class AppendDatabaseAction implements BaseAction {
     }
 
     private static void mergeFromBibtex(JabRefFrame frame, BasePanel panel, ParserResult pr, boolean importEntries,
-            boolean importStrings, boolean importGroups, boolean importSelectorWords) throws KeyCollisionException {
+            boolean importStrings, boolean importGroups) throws KeyCollisionException {
 
         BibDatabase fromDatabase = pr.getDatabase();
         List<BibEntry> appendedEntries = new ArrayList<>();
@@ -154,14 +152,6 @@ public class AppendDatabaseAction implements BaseAction {
                 // required here
                 frame.getGroupSelector().addGroups(newGroups, ce);
             });
-        }
-
-        if (importSelectorWords) {
-            for (String s : meta) {
-                if (s.startsWith(MetaData.SELECTOR_META_PREFIX)) {
-                    panel.getBibDatabaseContext().getMetaData().putData(s, meta.getData(s));
-                }
-            }
         }
 
         ce.end();